+++ Knob.js
@@ -28,6 +28,7 @@ class Knob extends React.Component {
     displayInput: React.PropTypes.bool,
     angleArc: React.PropTypes.number,
     angleOffset: React.PropTypes.number,
+    disableMouseWheel: React.PropTypes.bool,
   };
 
   static defaultProps = {
@@ -52,6 +53,7 @@ class Knob extends React.Component {
     displayInput: true,
     angleArc: 360,
     angleOffset: 0,
+    disableMouseWheel: false
   };
 
   constructor(props) {
@@ -281,7 +283,7 @@ class Knob extends React.Component {
   render = () => (
     <div
       style={{ width: this.w, height: this.h, display: 'inline-block' }}
-      onWheel={this.props.readOnly ? null : this.handleWheel}
+      onWheel={this.props.readOnly || this.props.disableMouseWheel ? null : this.handleWheel}
     >
       <canvas
         ref={(ref) => { this.canvasRef = ref; }}