+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.4.0
 * Added `displayCustom`, which allows ability to render custom component in the centre of the knob, if `displayInput` is set to false.
 * Added prop to disable mouse wheel action.
 * Added prop to add title attribute on the wheel.